+++ react/components/OrganizationUsersTable.tsx
@@ -522,6 +522,7 @@ const OrganizationUsersTable: FunctionComponent<Props> = ({
   const handleInputSearchSubmit = () => {
     refetch({
       ...variableState,
+      organizationId: isSalesAdmin ? null : organizationId,
       page: 1,
     })
   }