+++ react/admin/CostCenterDetails.tsx
@@ -1,4 +1,3 @@
-/* eslint-disable no-console */
 import type { FunctionComponent } from 'react'
 import React, { useEffect, useState } from 'react'
 import { useQuery, useMutation, useLazyQuery } from 'react-apollo'
@@ -31,16 +30,22 @@ import {
 import { StyleguideInput } from 'vtex.address-form/inputs'
 import { addValidation } from 'vtex.address-form/helpers'
 
-import { costCenterMessages as messages } from './utils/messages'
+import {
+  costCenterMessages as messages,
+  organizationCustomFieldsMessages as orgaizationMessages,
+} from './utils/messages'
 import { setGUID, getEmptyAddress, isValidAddress } from '../utils/addresses'
 import { validatePhoneNumber } from '../modules/formValidators'
 import GET_COST_CENTER from '../graphql/getCostCenter.graphql'
 import GET_ORGANIZATION from '../graphql/getOrganization.graphql'
 import UPDATE_COST_CENTER from '../graphql/updateCostCenter.graphql'
 import SET_MARKETING_TAGS from '../graphql/setMarketingTags.graphql'
+import GET_MARKETING_TAGS from '../graphql/getMarketingTags.graphql'
 import DELETE_COST_CENTER from '../graphql/deleteCostCenter.graphql'
 import GET_LOGISTICS from '../graphql/getLogistics.graphql'
-import GET_MARKETING_TAGS from '../graphql/getMarketingTags.graphql'
+import GET_B2B_CUSTOM_FIELDS from '../graphql/getB2BCustomFields.graphql'
+import { joinById } from './OrganizationDetails'
+import CustomFieldInput from './OrganizationDetailsCustomField'
 
 const CostCenterDetails: FunctionComponent = () => {
   const { formatMessage } = useIntl()
@@ -85,6 +90,8 @@ const CostCenterDetails: FunctionComponent = () => {
     addValidation(getEmptyAddress(country))
   )
 
+  const [customFieldsState, setCustomFieldsState] = useState<CustomField[]>([])
+
   const [tags, setTags] = useState([] as string[])
   const [tagName, setTagName] = useState('')
 
@@ -93,6 +100,12 @@ const CostCenterDetails: FunctionComponent = () => {
     skip: !params?.id,
     ssr: false,
     fetchPolicy: 'network-only',
+  })
+
+  const { data: defaultCustomFieldsData } = useQuery(GET_B2B_CUSTOM_FIELDS, {
+    variables: { id: params?.id },
+    skip: !params?.id,
+    ssr: false,
   })
 
   const [getOrganization, { data: organizationData }] = useLazyQuery(
@@ -195,6 +208,7 @@ const CostCenterDetails: FunctionComponent = () => {
         }),
         phoneNumber,
         businessDocument,
+        customFields: customFieldsState,
         stateRegistration,
       },
     }
@@ -206,7 +220,7 @@ const CostCenterDetails: FunctionComponent = () => {
     updateCostCenter({ variables })
       .then(() => {
         showToast({
-          type: 'success',
+          variant: 'positive',
           message: formatMessage(messages.toastUpdateSuccess),
         })
         refetch()
@@ -216,7 +230,7 @@ const CostCenterDetails: FunctionComponent = () => {
       .catch(error => {
         console.error(error)
         showToast({
-          type: 'error',
+          variant: 'critical',
           message: formatMessage(messages.toastUpdateFailure),
         })
         setLoadingState(false)
@@ -233,7 +247,7 @@ const CostCenterDetails: FunctionComponent = () => {
       .catch(error => {
         console.error(error)
         showToast({
-          type: 'error',
+          variant: 'critical',
           message: formatMessage(messages.toastDeleteFailure),
         })
         setLoadingState(false)
@@ -329,7 +343,7 @@ const CostCenterDetails: FunctionComponent = () => {
       handleCloseModals()
     } else {
       showToast({
-        type: 'error',
+        variant: 'critical',
         message: formatMessage(messages.duplicateAddress),
       })
     }
@@ -404,6 +418,30 @@ const CostCenterDetails: FunctionComponent = () => {
       onClick: () => handleDeleteAddressModal(addressId),
     },
   ]
+
+  // CostCenter custom fields
+
+  useEffect(() => {
+    const customFieldsToShow = joinById(
+      data?.getCostCenterById?.customFields || [],
+      defaultCustomFieldsData?.getB2BSettings.costCenterCustomFields || []
+    )
+
+    setCustomFieldsState(customFieldsToShow)
+  }, [
+    data?.getCostCenterById?.customFields &&
+      defaultCustomFieldsData?.getB2BSettings.costCenterCustomFields,
+  ])
+
+  const handleCustomFieldsUpdate = (
+    index: number,
+    customField: CustomField
+  ) => {
+    const newCustomFields = [...customFieldsState]
+
+    newCustomFields[index] = customField
+    setCustomFieldsState(newCustomFields)
+  }
 
   if (!data) {
     return (
@@ -602,6 +640,17 @@ const CostCenterDetails: FunctionComponent = () => {
             </Card>
           </div>
         </div>
+      </PageBlock>
+
+      <PageBlock title={formatMessage(orgaizationMessages.customFieldsTitle)}>
+        {customFieldsState?.map((customField: CustomField, index: number) => (
+          <CustomFieldInput
+            key={`${customField.name}`}
+            customField={customField}
+            index={index}
+            handleUpdate={handleCustomFieldsUpdate}
+          />
+        ))}
       </PageBlock>
 
       {/* New Address Modal */}